A more advanced configuration of circleci
This repo was a little rusty and needed to be dusted off. To get tests
going I had to update the package.json file to use a bit more modern of
a set of deps, add a linting rc file, and clean up the lint errors. This
all runs in CircleCi now.